requirements.txt
albumentations==1.2.0
numpy==1.21.2
scipy==1.7.3
torch==1.10.1
torchvision==0.11.2

pyyaml~=6.0